Rename source code repository module
